requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
newspaper3k==0.2.8
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
sumy==0.11.0
python-dotenv==1.0.0
openai==1.3.7